+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/AppDelegate.m
@@ -1,4 +1,5 @@
 #import "AppDelegate.h"
+@import CoreData;
 @import Sentry;
 
 @interface