+++ Samples/iOS-Swift/iOS-Swift/ViewController.swift
@@ -4,6 +4,8 @@ import UIKit
 class ViewController: UIViewController {
     
     @IBOutlet weak var dsnTextField: UITextField!
+    @IBOutlet weak var anrFullyBlockingButton: UIButton!
+    @IBOutlet weak var anrFillingRunLoopButton: UIButton!
     
     private let dispatchQueue = DispatchQueue(label: "ViewController")
 
@@ -49,7 +51,7 @@ class ViewController: UIViewController {
         SentrySDK.addBreadcrumb(crumb: crumb)
     }
     
-    @IBAction func captureMessage(_ sender: Any) {        
+    @IBAction func captureMessage(_ sender: Any) {
         let eventId = SentrySDK.capture(message: "Yeah captured a message")
         // Returns eventId in case of successfull processed event
         // otherwise nil
@@ -149,6 +151,40 @@ class ViewController: UIViewController {
         }
     }
 
+    @IBAction func anrFullyBlocking(_ sender: Any) {
+        let buttonTitle = self.anrFullyBlockingButton.currentTitle
+        var i = 0
+        
+        for _ in 0...5_000_000 {
+            i += Int.random(in: 0...10)
+            i -= 1
+            
+            self.anrFullyBlockingButton.setTitle("\(i)", for: .normal)
+        }
+        
+        self.anrFullyBlockingButton.setTitle(buttonTitle, for: .normal)
+    }
+    
+    @IBAction func anrFillingRunLoop(_ sender: Any) {
+        let buttonTitle = self.anrFillingRunLoopButton.currentTitle
+        var i = 0
+
+        dispatchQueue.async {
+            for _ in 0...100_000 {
+                i += Int.random(in: 0...10)
+                i -= 1
+                
+                DispatchQueue.main.async {
+                    self.anrFillingRunLoopButton.setTitle("Work in Progress \(i)", for: .normal)
+                }
+            }
+            
+            DispatchQueue.main.async {
+                self.anrFillingRunLoopButton.setTitle(buttonTitle, for: .normal)
+            }
+        }
+    }
+    
     @IBAction func dsnChanged(_ sender: UITextField) {
         let options = Options()
         options.dsn = sender.text
@@ -186,6 +222,12 @@ class ViewController: UIViewController {
     @IBAction func showTableViewController(_ sender: Any) {
         let controller = TableViewController(style: .plain)
         controller.title = "Table View Controller"
+        navigationController?.pushViewController(controller, animated: false)
+    }
+    
+    @IBAction func useCoreData(_ sender: Any) {
+        let controller = CoreDataViewController()
+        controller.title = "CoreData"
         navigationController?.pushViewController(controller, animated: false)
     }
 }